Fixes Crash when client had bad password
We were deleting the client while in a function call-chain created by the client. This caused invalid read/write operation.
This has now been fix by adding more throw to make the "BadPassword" exception go up the call-chain until being higher than the client to delete in the stack, and then deleting it